# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  asset-registration-management-system_backend_internal_model.User:
    properties:
      adDn:
        type: string
      authSource:
        type: string
      createdAt:
        type: string
      department:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: integer
      lastLoginAt:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Role'
      status:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  internal_httpapi.adConfigRequest:
    properties:
      advancedFilter:
        type: boolean
      baseDn:
        type: string
      bindDn:
        type: string
      bindPassword:
        type: string
      enabled:
        type: boolean
      excludeDisabled:
        type: boolean
      filterUserObject:
        type: boolean
      ldapUrl:
        type: string
      loginAttribute:
        type: string
      userFilter:
        type: string
    required:
    - baseDn
    - bindDn
    - ldapUrl
    type: object
  internal_httpapi.adImportRequest:
    properties:
      role:
        $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Role'
      status:
        type: string
      username:
        type: string
    required:
    - username
    type: object
...
  internal_httpapi.errorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: 企业内部服务器资产台账与工单流程管理系统 REST API。
  title: Asset Registration Management System API
  version: "1.0"
paths:
  /ad/config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: AD 配置详情
      tags:
      - settings
    put:
      consumes:
      - application/json
      parameters:
      - description: AD 配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.adConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 保存 AD 配置
      tags:
      - settings
  /ad/import-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 导入参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.adImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 导入 AD 用户
      tags:
      - settings
  /ad/lookup-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.adLookupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_service.ADUserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 查询 AD 用户
      tags:
      - settings
  /ad/test:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 测试 AD 连接
      tags:
      - settings
  /assets:
    get:
      parameters:
      - description: 关键词
        in: query
        name: q
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 资产列表
      tags:
      - assets
    post:
      consumes:
      - application/json
      parameters:
      - description: 资产信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.assetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Asset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建资产
      tags:
      - assets
  /assets/{id}:
    delete:
      parameters:
      - description: 资产 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除资产
      tags:
      - assets
    get:
      parameters:
      - description: 资产 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Asset'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 资产详情
      tags:
      - assets
    put:
      consumes:
      - application/json
      parameters:
      - description: 资产 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 资产信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.assetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Asset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新资产
      tags:
      - assets
  /assets/{id}/history:
    get:
      parameters:
      - description: 资产 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/asset-registration-management-system_backend_internal_model.AssetSnapshot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 资产变更历史
      tags:
      - assets
  /assets/export:
    get:
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出资产 CSV
      tags:
      - assets
  /assets/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV 或 XLSX 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 导入资产
      tags:
      - assets
  /assets/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 资产统计
      tags:
      - assets
  /assets/template:
    get:
      parameters:
      - description: 模板格式 csv 或 xlsx
        in: query
        name: format
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 下载资产导入模板
      tags:
      - assets
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      summary: 用户登录
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 用户退出
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 当前用户
      tags:
      - auth
  /discovery/rules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/asset-registration-management-system_backend_internal_model.DiscoveryRule'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 发现规则列表
      tags:
      - discovery
    post:
      consumes:
      - application/json
      parameters:
      - description: 发现规则
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.discoveryRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.DiscoveryRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建发现规则
      tags:
      - discovery
  /discovery/rules/{id}:
    delete:
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除发现规则
      tags:
      - discovery
    put:
      consumes:
      - application/json
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 发现规则
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.discoveryRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.DiscoveryRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新发现规则
      tags:
      - discovery
  /discovery/rules/{id}/run:
    post:
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.DiscoveryRun'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 手动触发发现任务
      tags:
      - discovery
  /discovery/rules/{id}/test:
    post:
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 试跑发现规则
      tags:
      - discovery
  /discovery/runs:
    get:
      parameters:
      - description: 规则 ID
        in: query
        name: ruleId
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 运行记录列表
      tags:
      - discovery
  /discovery/runs/{id}:
    get:
      parameters:
      - description: 运行 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.DiscoveryRun'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 运行记录详情
      tags:
      - discovery
  /discovery/runs/{id}/adopt:
    post:
      consumes:
      - application/json
      parameters:
      - description: 运行 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 主机 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.discoveryHostActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 纳管新发现主机
      tags:
      - discovery
  /discovery/runs/{id}/apply:
    post:
      consumes:
      - application/json
      parameters:
      - description: 运行 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 主机 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.discoveryHostActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 应用发现变更到资产
      tags:
      - discovery
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - users
  /settings/mail:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 邮件配置详情
      tags:
      - settings
    put:
      consumes:
      - application/json
      parameters:
      - description: 邮件配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.mailConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 保存邮件配置
      tags:
      - settings
  /settings/mail/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: 测试参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.mailTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 测试邮件发送
      tags:
      - settings
  /tickets:
    get:
      parameters:
      - description: 视图 todo/submitted/all
        in: query
        name: view
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 工单列表
      tags:
      - tickets
    post:
      consumes:
      - application/json
      parameters:
      - description: 工单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.ticketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Ticket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建工单
      tags:
      - tickets
  /tickets/{id}:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Ticket'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 工单详情
      tags:
      - tickets
    put:
      consumes:
      - application/json
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 工单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.ticketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.Ticket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新工单
      tags:
      - tickets
...
  /tickets/{id}/archive/download:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 下载工单归档 PDF
      tags:
      - tickets
  /tickets/{id}/attachments:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 工单附件列表
      tags:
      - tickets
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 附件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.TicketAttachment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 上传工单附件
      tags:
      - tickets
  /tickets/{id}/attachments/{attachmentId}/download:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 附件 ID
        in: path
        name: attachmentId
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 下载工单附件
      tags:
      - tickets
...
  /tickets/{id}/comments:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 工单评论列表
      tags:
      - tickets
    post:
      consumes:
      - application/json
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.ticketCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.TicketComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建工单评论
      tags:
      - tickets
...
  /tickets/archives/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工单 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.ticketArchiveBatchRequest'
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 批量下载工单归档 ZIP
      tags:
      - tickets
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.userRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - users
  /users/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.userRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - users
  /workflows:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 流程配置列表
      tags:
      - workflows
  /workflows/{type}:
    get:
      parameters:
      - description: 工单类型
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.TicketWorkflow'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 工单类型流程配置
      tags:
      - workflows
    put:
      consumes:
      - application/json
      parameters:
      - description: 工单类型
        in: path
        name: type
        required: true
        type: string
      - description: 流程配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.workflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.TicketWorkflow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 保存工单类型流程配置
      tags:
      - workflows
  /workflows/{type}/enable:
    post:
      parameters:
      - description: 工单类型
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset-registration-management-system_backend_internal_model.TicketWorkflow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: 启用工单类型流程
      tags:
      - workflows
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"