# backend/internal/httpapi/router.go
package httpapi

import (
	"net/http"
	"strings"
	"time"

	_ "asset-registration-management-system/backend/docs"
	"asset-registration-management-system/backend/internal/config"
	"asset-registration-management-system/backend/internal/model"
	"asset-registration-management-system/backend/internal/service"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	"gorm.io/gorm"
)

type Dependencies struct {
	Config   config.Config
	DB       *gorm.DB
	Roles    []model.Role
	AD       service.ADClient
	Archiver service.TicketArchiver
	Mail     service.MailSender
}

func NewRouter(dep Dependencies) *gin.Engine {
	router := gin.New()
	router.Use(gin.Logger(), gin.Recovery(), cors(dep.Config.CORS.AllowedOrigins))

	adClient := dep.AD
	if adClient == nil {
		adClient = service.LDAPADClient{}
	}
	archiver := dep.Archiver
	if archiver == nil {
		archiver = service.LibreOfficeTicketArchiver{}
	}
	mailSender := dep.Mail
	if mailSender == nil {
		mailSender = service.SMTPMailSender{}
	}
	h := NewHandler(dep.Config, dep.DB, dep.Roles, adClient, archiver, mailSender)

	router.GET("/healthz", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{"status": "ok", "time": time.Now().UTC()})
	})
	if dep.Config.Swagger.Enabled {
		router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	api := router.Group("/api/v1")
	api.POST("/auth/login", h.Login)
	api.POST("/auth/logout", h.Logout)
	api.GET("/auth/me", h.AuthRequired(), h.Me)

	api.GET("/roles", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin), h.ListRoles)

	ad := api.Group("/ad", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin))
	ad.GET("/config", h.GetADConfig)
	ad.PUT("/config", h.SaveADConfig)
	ad.POST("/test", h.TestADConnection)
	ad.POST("/lookup-user", h.LookupADUser)
	ad.POST("/import-user", h.ImportADUser)

	settings := api.Group("/settings", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin))
	settings.GET("/mail", h.GetMailConfig)
	settings.PUT("/mail", h.SaveMailConfig)
	settings.POST("/mail/test", h.TestMailConfig)

	users := api.Group("/users", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin))
	users.GET("", h.ListUsers)
	users.POST("", h.CreateUser)
	users.PUT("/:id", h.UpdateUser)

	typeApprovers := api.Group("/ticket-type-approvers", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin))
	typeApprovers.GET("", h.ListTicketTypeApprovers)
	typeApprovers.PUT("/:type", h.SetTicketTypeApprover)

	workflows := api.Group("/workflows", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin))
	workflows.GET("", h.ListWorkflows)
	workflows.GET("/:type", h.GetWorkflow)
	workflows.PUT("/:type", h.SaveWorkflow)
	workflows.POST("/:type/enable", h.EnableWorkflow)

	assets := api.Group("/assets", h.AuthRequired())
	assets.GET("", h.ListAssets)
	assets.GET("/stats", h.AssetStats)
	assets.POST("", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.CreateAsset)
	assets.GET("/export", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.ExportAssets)
	assets.GET("/template", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.DownloadAssetImportTemplate)
	assets.POST("/import", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.ImportAssets)
	assets.GET("/:id", h.GetAsset)
	assets.GET("/:id/history", h.ListAssetHistory)
	assets.PUT("/:id", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.UpdateAsset)
	assets.DELETE("/:id", h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager), h.DeleteAsset)

	discovery := api.Group("/discovery", h.AuthRequired(), h.RequireAnyRole(model.RoleAdmin, model.RoleAssetManager))
	discovery.GET("/rules", h.ListDiscoveryRules)
	discovery.POST("/rules", h.CreateDiscoveryRule)
	discovery.PUT("/rules/:id", h.UpdateDiscoveryRule)
	discovery.DELETE("/rules/:id", h.DeleteDiscoveryRule)
	discovery.POST("/rules/:id/run", h.StartDiscoveryRun)
	discovery.POST("/rules/:id/test", h.TestDiscoveryRun)
	discovery.GET("/runs", h.ListDiscoveryRuns)
	discovery.GET("/runs/:id", h.GetDiscoveryRun)
	discovery.POST("/runs/:id/adopt", h.AdoptDiscoveryHosts)
	discovery.POST("/runs/:id/apply", h.ApplyDiscoveryHosts)

	tickets := api.Group("/tickets", h.AuthRequired())
	tickets.GET("", h.ListTickets)
	tickets.POST("", h.CreateTicket)
	tickets.POST("/archives/download", h.DownloadTicketArchives)
	tickets.GET("/:id", h.GetTicket)
	tickets.PUT("/:id", h.UpdateTicket)
	tickets.GET("/:id/comments", h.ListTicketComments)
	tickets.POST("/:id/comments", h.CreateTicketComment)
	tickets.GET("/:id/attachments", h.ListTicketAttachments)
	tickets.POST("/:id/attachments", h.UploadTicketAttachment)
	tickets.GET("/:id/attachments/:attachmentId/download", h.DownloadTicketAttachment)
	tickets.GET("/:id/archive/download", h.DownloadTicketArchive)
	for _, action := range []string{"submit", "approve", "reject", "start", "complete", "accept", "cancel"} {
		tickets.POST("/:id/"+action, h.TicketAction(action))
	}

	return router
}

func cors(allowedOrigins string) gin.HandlerFunc {
	return func(c *gin.Context) {
		origin := c.Request.Header.Get("Origin")
		allow := allowedOrigins == "*" || originAllowed(origin, allowedOrigins)
		if allow {
			if allowedOrigins == "*" {
				c.Header("Access-Control-Allow-Origin", "*")
			} else {
				c.Header("Access-Control-Allow-Origin", origin)
			}
			c.Header("Access-Control-Allow-Headers", "Authorization, Content-Type")
			c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		}
		if c.Request.Method == http.MethodOptions {
			c.AbortWithStatus(http.StatusNoContent)
			return
		}
		c.Next()
	}
}

func originAllowed(origin, allowedOrigins string) bool {
	for _, item := range strings.Split(allowedOrigins, ",") {
		if strings.TrimSpace(item) == origin {
			return true
		}
	}
	return false
}